+++ packages/ui/src/hooks/system/index.ts
@@ -1,3 +1,0 @@
-import useSystemRoute from './useSystemRoute';
-
-export { useSystemRoute };