+++ shared/framework-kernel/src/lib/hooks/index.ts
@@ -1,6 +1,7 @@
 import useEditFinish from './useEditFinish';
 import usePasskey from './usePasskey';
 import useSystemMenu from './useSystemMenu';
+import useSystemRoute from './useSystemRoute';
 import useSystemTheme from './useSystemTheme';
 
-export { useEditFinish, usePasskey, useSystemMenu, useSystemTheme };
+export { useEditFinish, usePasskey, useSystemMenu, useSystemRoute, useSystemTheme };