+++ shared/oss-apis/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @herodotus-cloud/oss-apis
 
+## 1.7.11
+
+### Patch Changes
+
+- 框架代码迁移至模块主体完成
+- Updated dependencies
+  - @herodotus-cloud/core@1.7.11
+
 ## 1.7.10
 
 ### Patch Changes